floating icons
- add new search and add icons for future features
