+++ src/Pages/homepage/allRecipes/AllRecipes.jsx
@@ -1,12 +1,14 @@
 import React from "react";
 import T from "i18n-react";
 import PropTypes from "prop-types";
+import Zoom from '@material-ui/core/Zoom';
+import Search from '@material-ui/icons/Search';
 import RecipeReviewCard from "Components/recipeCard/recipeCard";
 import { Container, Row, Col } from 'react-grid-system';
 import CardLoader from "Components/loaders/cardLoader/CardLoader";
 import NoRecipes from "Components/placeholders/noRecipes/NoRecipes";
 
-import { CenteredRow, Title, ContainerStyled } from "./styles";
+import { CenteredRow, Title, ContainerStyled, StyledFab } from "./styles";
 
 const AllRecipes = ({ recipes }) => {
   return (
@@ -33,6 +35,17 @@ const AllRecipes = ({ recipes }) => {
           )}
         </CenteredRow>
       </Container>
+      {recipes && (
+        <Zoom
+          in={!!recipes}
+          style={{ transitionDelay: "100ms" }}
+          unmountOnExit
+        >
+          <StyledFab color="secondary" aria-label="add">
+            <Search />
+          </StyledFab>
+        </Zoom>
+      )}
     </ContainerStyled>
   );
 };
