+++ src/Pages/homepage/allRecipes/styles.js
@@ -2,6 +2,7 @@
 import styled from "styled-components";
 import { Row } from 'react-grid-system';
 import Typography from '@material-ui/core/Typography';
+import Fab from '@material-ui/core/Fab';
 
 export const CenteredRow = styled(Row)`
   &&& {
@@ -27,5 +28,17 @@ export const ContainerStyled = styled.div`
     height: 85%;
     overflow: scroll;
     margin-bottom: 50px;
+  }
+`;
+
+export const StyledFab = styled(Fab)`
+  &&& {
+    position: fixed;
+    bottom: 85px;
+    right: 30px;
+    cursor: pointer;
+    @media only screen and (min-width: 768px) {
+      position: absolute;
+    }
   }
 `;
