+++ src/Pages/homepage/favorites/styles.js
@@ -2,6 +2,7 @@
 import styled from "styled-components";
 import { Row } from 'react-grid-system';
 import Typography from '@material-ui/core/Typography';
+import Fab from '@material-ui/core/Fab';
 
 export const CenteredRow = styled(Row)`
   &&& {
@@ -25,4 +26,16 @@ export const ContainerStyled = styled.div`
   width: 100%;
   height: 85%;
   overflow: scroll;
+`;
+
+export const StyledFab = styled(Fab)`
+  &&& {
+    position: fixed;
+    bottom: 85px;
+    right: 30px;
+    cursor: pointer;
+    @media only screen and (min-width: 768px) {
+      position: absolute;
+    }
+  }
 `;
